Restrict permitted resource requests
In the last commit, any and all resource requests were permitted. This was
simply an attempt to make the static (Angular-generated) resources fetchable
in any way possible, so that the site could be used. Of course, the downside
is that removes the authentification provided by the JWT system entirely.

After researching, it seems the only solution online within Spring is to
list every resource individually.

Signed-off-by: Philippa Hack <philippa.hack@tngtech.com>