Fix tests
Tests were failing because the Accept RequestParameter on the
findVersion method had been added since the tests were written, and no
default value was supplied, so the tests failed to run.

Adds default values, however the unit test and integration test for the
version controller require the default specified in two different ways
due to how the findVersion function is used in each.

Signed-off-by: Philippa Hack <philippa.hack@tngtech.com>