Fix mess in ordering
By replacing the dictionary that used to map the fileType string to
download functions, with a function which essentially does the same
thing, the difficulties with the lambda functions go away (eg. the
linter considered the functions as class attributes not methods, and
would not hoist the definitions. It was impossible to make the code work
without disabling the linter).

Signed-off-by: Philippa Hack <philippa.hack@tngtech.com>