Correct specification route
After clicking around the website, it seems that there was never an
Angular route configured for 'specifications/:specificationId'. However in
the specification search, this was the router link used when a user
clicked on a specification matching the search string. As Angular had no
route configured, clicking just returned the webapp to the home route
('/').

This change replaces that non-existent route with the page of the most
recent version of the specification.

Signed-off-by: Philippa Hack <philippa.hack@tngtech.com>