Permit GET requests to non-root URLs on monoBuild
Angular routes are usually not translated to HTTP requests. When running
off ng serve, Angular will correctly interpret a GET request to, say,
"http://localhost:8080/specifications", but when the Angular files are
served statically, only requests to the root path are possible, and any
subsequent URLs are "mocked" in the browser address bar. This is useful if
a user ever refreshes the page, or sends a URL to someone, and doesn't
want to get a 404 response.

However in Spring it is possible to route any request through to Angular,
like so.

Signed-off-by: Philippa Hack <philippa.hack@tngtech.com>