Change API endpoint and add 404 handlers
This is two feature changes that unfortunately got combined.

Changes the backend request mappings to reflect the API redefinition.

Changes the findOne methods for Specifications and Versions to support if
the requested version is not found. Before, they assumed a find operation
was always successful, and the server threw an exception if otherwise,
returning 500. Now API calls return 404.

Signed-off-by: Philippa Hack <philippa.hack@tngtech.com>