Render API GUI based on language
Both SwaggerUI and GraphiQL wrapper components need to have the same
Angular route, and this solution has them as child components under a
generic Version View.

So that the file format can be identified when the client fetches the
data, this adds a language field to the Version entity, which is set on
file upload based on the extension.

However although the SwaggerUI appears, it is not interactive.

Signed-off-by: Philippa Hack <philippa.hack@tngtech.com>