Add Flyway migration
With Flyway, we have an explicit database schema that looks sensible. We
do not depend anymore on Hibernate doing something good.
For this reason, the length of fields has been set to something
reasonable, and the names are not screaming anymore.

This also fixes some bugs with the tests:
* Both production and tests added some values to the DB, but they used
the same IDs. For now, production has been commented out and replaced
with a useless query, as it must not be empty.
* The tests expected different values than were provided.

Signed-off-by: Fabian Winter <fabian.winter@tngtech.com>