Fix DB relationship
Hibernate is confusing sometimes, and in this case, the problem was that
`@OneToMany`'s `mappedBy` refers not to the *table*, but to the *field*
in the other entity.

This was even more malicious here because the `version` table has a
field `version`. While the authors wanted to refer to the table,
Hibernate took that as a reference to the field - which makes no sense
at all from the outside.
And then, the version field is a `String`, but the ID of the reference
is `Bytes` (because it is a UUID). That leaves us with a very, very
unhelpful error and lots of debugging and crying.

Signed-off-by: Fabian Winter <fabian.winter@tngtech.com>